Extract menu states to utils
