+++ packages/components/src/tools-panel/utils.js
@@ -1,0 +1,5 @@
+export const MENU_STATES = {
+	CHECKED: 'checked',
+	UNCHECKED: 'unchecked',
+	DISABLED: 'disabled',
+};
